Force new manager resource when changing type
The API started misbehaving when changing type and setting properties
from the previous type to null.

The API client is very simple and always sends all of the possible
properties for all manager types.

Forcing the creation of a new manager when changing the type avoids
these edge cases and inconsistencies, keeping the API client simple.